.vercel/cache/main/requirements.txt
fastapi
pymongo
passlib
pydantic[email]
python-multipart
python-decouple